+++ src/test/java/christmas/domain/event/SpecialDiscountTest.java
@@ -1,39 +1,44 @@
-package christmas.domain;
+package christmas.domain.event;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import christmas.domain.Date;
+import christmas.domain.Order;
+import christmas.domain.event.SpecialDiscount;
 import christmas.domain.menu.Menu;
 import java.util.EnumMap;
+import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class WeekdayDiscountTest {
-    WeekdayDiscount weekdayDiscount = new WeekdayDiscount();
-    Date weekendDate = new Date(1);
-    Date weekdayDate = new Date(25);
+class SpecialDiscountTest {
+    SpecialDiscount specialDiscount = new SpecialDiscount();
+    List<Integer> starredDates = List.of(3, 10, 17, 24, 25, 31);
 
     @DisplayName("10000원 미만인 경우 할인 적용되지 않는다.")
     @Test
     void notApplicableOrder() {
+
         Map<Menu, Integer> menus = new EnumMap<Menu, Integer>(Menu.class);
         menus.put(Menu.fromDescription("타파스"), 1);
 
-        assertThat(weekdayDiscount.canBeApplied(weekdayDate, new Order(menus)))
+        assertThat(specialDiscount.canBeApplied(new Date(starredDates.stream().findAny().get()), new Order(menus)))
                 .isEqualTo(false);
     }
 
-    @DisplayName("평일이 아니면 적용되지 않는다.")
+    @DisplayName("정해진 날짜가 아니면 할인이 적용되지 않는다.")
     @Test
     void notApplicableDate() {
-        assertThat(weekdayDiscount.canBeApplied(weekdayDate, createOrder())).isEqualTo(true);
-        assertThat(weekdayDiscount.canBeApplied(weekendDate, createOrder())).isEqualTo(false);
+        assertThat(specialDiscount.canBeApplied(new Date(3), createOrder())).isEqualTo(true);
+        assertThat(specialDiscount.canBeApplied(new Date(1), createOrder())).isEqualTo(false);
     }
 
     @DisplayName("할인 금액 테스트")
     @Test
     void amount() {
-        assertThat(weekdayDiscount.amount(weekdayDate, createOrder())).isEqualTo(-4046);
+        assertThat(specialDiscount.amount(new Date(starredDates.stream().findAny().get()), createOrder()))
+                .isEqualTo(-1000);
     }
 
     Order createOrder() {
@@ -46,4 +51,5 @@ class WeekdayDiscountTest {
         menus.put(Menu.fromDescription("샴페인"), 1);
         return new Order(menus);
     }
+
 }
